Separting the action test into own suite
